@@ -4,6 +4,7 @@ import java.io.*;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
+import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.http.*;
@@ -11,9 +12,9 @@ import jakarta.servlet.annotation.*;
 
 @WebServlet(name = "Ticket", value="/Ticket")
 @MultipartConfig(fileSizeThreshold = 5_242_880, maxFileSize = 20_971_520L, maxRequestSize = 41_943_040L)
-public class TicketServlet extends HttpServlet {
-    private volatile int ticketId = 1;
-    private Map<Integer, Ticket> ticketDB = new LinkedHashMap<>();
+public class TicketServlet extends HttpServlet{
+    private volatile int Ticket_ID = 1;
+    private Map<Integer, Ticket> TicketDB = new LinkedHashMap<>();
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -25,18 +26,12 @@ public class TicketServlet extends HttpServlet {
             action = "list";
         }
         switch(action) {
-            case "createTicket":
-                showPostForm(request, response);
-                break;
-            case "view":
-                viewTicket(request, response);
-                break;
-            case "download":
-                downloadAttachment(request, response);
-                break;
-            default:
-                listTickets(request, response);
+            case "createTicket" -> showPostForm(request, response);
+            case "view" -> viewPost(request, response);
+            case "download" -> downloadImage(request, response);
+            default -> listPosts(request, response); // this the list and any other
         }
+
     }
 
     @Override
@@ -49,34 +44,18 @@ public class TicketServlet extends HttpServlet {
             action = "list";
         }
         switch(action) {
-            case "create":
-                createTicket(request, response);
-                break;
-            default:
-                response.sendRedirect("Ticket");
+            case "create" -> createPost(request, response);
+            default -> response.sendRedirect("Ticket"); // this the list and any other
         }
     }
 
-    private void listTickets(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        PrintWriter out = response.getWriter();
-
-        out.println("<html><body><h2>Ticket Posts</h2>");
-        out.println("<a href=\"Ticket?action=createTicket\">Create Post</a><br><br>");
-
-        if (ticketDB.isEmpty()) {
-            out.println("There are no Ticket posts yet...");
-        } else {
-            for (int id : ticketDB.keySet()) {
-                Ticket ticket = ticketDB.get(id);
-                out.println("Ticket #" + id);
-                out.println(": <a href=\"Ticket?action=view&TicketId=" + id + "\">");
-                out.println(ticket.getTitle() + "</a><br>");
-            }
-        }
-        out.println("</body></html>");
+    private void listPosts(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+        request.setAttribute("ticketDB", TicketDB);
+        RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/view/listTickets.jsp");
+        dispatcher.forward(request, response);
     }
 
-    private void createTicket(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    private void createPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         Ticket ticket = new Ticket();
         ticket.setTitle(request.getParameter("title"));
         ticket.setDate();
@@ -84,7 +63,7 @@ public class TicketServlet extends HttpServlet {
 
         Part file = request.getPart("file1");
         if (file != null) {
-            Image image = processImage(file);
+            Image image = this.processImage(file);
             if (image != null) {
                 ticket.setImage(image);
             }
@@ -92,14 +71,14 @@ public class TicketServlet extends HttpServlet {
 
         int id;
         synchronized(this) {
-            id = this.ticketId++;
-            ticketDB.put(id, ticket);
+            id = this.Ticket_ID++;
+            TicketDB.put(id, ticket);
         }
 
         response.sendRedirect("Ticket?action=view&TicketId=" + id);
     }
 
-    private Image processImage(Part file) throws IOException {
+    private Image processImage(Part file) throws IOException{
         InputStream in = file.getInputStream();
         ByteArrayOutputStream out = new ByteArrayOutputStream();
 
@@ -109,14 +88,14 @@ public class TicketServlet extends HttpServlet {
             out.write(bytes, 0, read);
         }
 
-        Image image = new Image();
+        com.example.craytonmacholanassignment42.Image image = new com.example.craytonmacholanassignment42.Image();
         image.setName(file.getSubmittedFileName());
         image.setContents(out.toByteArray());
 
         return image;
     }
 
-    private void downloadAttachment(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    private void downloadImage(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
         String idString = request.getParameter("TicketId");
 
         Ticket ticket = getTicket(idString, response);
@@ -139,42 +118,25 @@ public class TicketServlet extends HttpServlet {
         out.write(image.getContents());
     }
 
-    private void viewTicket(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    private void viewPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
         String idString = request.getParameter("TicketId");
 
         Ticket ticket = getTicket(idString, response);
 
-        PrintWriter out = response.getWriter();
-        out.println("<html><body><h2>Ticket Post</h2>");
-        out.println("<h3>" + ticket.getTitle()+ "</h3>");
-        out.println("<p>Date: " + ticket.getDate() + "</p>");
-        out.println("<p>" + ticket.getBody() + "</p>");
-        if (ticket.hasImage()) {
-            out.println("<a href=\"Ticket?action=download&TicketId=" +
-                    idString + "&image="+ ticket.getImage().getName() + "\">" +
-                    ticket.getImage().getName() + "</a><br><br>");
+        if (ticket != null) {
+            request.setAttribute("ticketId", idString);
+            request.setAttribute("ticket", ticket);
+            RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/view/viewTicket.jsp");
+            dispatcher.forward(request, response);
         }
-        out.println("<a href=\"ticket\">Return to Ticket list</a>");
-        out.println("</body></html>");
     }
 
-    private void showPostForm(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        PrintWriter out = response.getWriter();
-
-        out.println("<html><body><h2>Create a Ticket Post</h2>");
-        out.println("<form method=\"POST\" action=\"Ticket\" enctype=\"multipart/form-data\">");
-        out.println("<input type=\"hidden\" name=\"action\" value=\"create\">");
-        out.println("Title:<br>");
-        out.println("<input type=\"text\" name=\"title\"><br><br>");
-        out.println("Body:<br>");
-        out.println("<textarea name=\"body\" rows=\"25\" cols=\"100\"></textarea><br><br>");
-        out.println("<b>Image</b><br>");
-        out.println("<input type=\"file\" name=\"file1\"><br><br>");
-        out.println("<input type=\"submit\" value=\"Submit\">");
-        out.println("</form></body></html>");
+    private void showPostForm(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException{
+        RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/jsp/view/ticketForm.jsp");
+        dispatcher.forward(request, response);
     }
 
-    private Ticket getTicket(String idString, HttpServletResponse response) throws ServletException, IOException {
+    private Ticket getTicket(String idString, HttpServletResponse response) throws ServletException, IOException{
         if (idString == null || idString.length() == 0) {
             response.sendRedirect("Ticket");
             return null;
@@ -182,15 +144,17 @@ public class TicketServlet extends HttpServlet {
 
         try {
             int id = Integer.parseInt(idString);
-            Ticket ticket = ticketDB.get(id);
+            Ticket ticket = TicketDB.get(id);
             if (ticket == null) {
                 response.sendRedirect("Ticket");
                 return null;
             }
             return ticket;
-        } catch(Exception e) {
+        }
+        catch(Exception e) {
             response.sendRedirect("Ticket");
             return null;
         }
     }
+
 }
